Downloads all songs on aws using DownloadAllSongsAndAlbums.sh
The issue was song names had to be specified in the download songs script to
download from aws. We want don't want to change the script everytime we
add new songs.

My solution was to create a new script that uses awscli to download all songs
from the aws solarcar onboard media bucket. The media player also displays
empty album art when it does not exist in the mp3 file.